用户登录，以及token登录初步添加

+++ src/main/java/com/msir/web/MenuController.java
@@ -4,6 +4,7 @@ import com.alibaba.fastjson.JSON;
 import com.msir.enums.MenuStateEnum;
 import com.msir.pojo.MenuDO;
 import com.msir.service.MenuService;
+import com.msir.utils.GlobalUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
@@ -11,6 +12,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.servlet.http.HttpServletResponse;
 import java.util.ArrayList;
 import java.util.List;
+
 
 /**
  * Created by Fantasy on 2017/5/28.
@@ -49,13 +51,8 @@ public class MenuController {
     @ResponseBody
     @RequestMapping(value = "/listMeta", method = RequestMethod.GET, produces = {"application/json; charset=utf-8"})
     public Object listMetaMenu(HttpServletResponse resp) {
-        if (resp.getStatus() == resp.SC_REQUEST_TIMEOUT) {
-            return new FinalResult<String>(
-                    true,
-                    "",
-                    "token失效",
-                    "菜单列表",
-                    "0");
+        if (GlobalUtils.authenErrorStatus(resp)) {
+            return GlobalUtils.authenErrorManager(resp);
         }
         FinalResult finalResult = new FinalResult<List>(
                 true,
@@ -73,7 +70,10 @@ public class MenuController {
      */
     @ResponseBody
     @RequestMapping(value = "/detail", method = RequestMethod.POST, produces = {"application/json; charset=utf-8"})
-    public Object saveMenu(MenuDO menuDO) {
+    public Object saveMenu(MenuDO menuDO, HttpServletResponse resp) {
+        if (GlobalUtils.authenErrorStatus(resp)) {
+            return GlobalUtils.authenErrorManager(resp);
+        }
         FinalResult finalResult;
         int menuServiceStatus = menuService.saveMenu(menuDO);
         if (menuServiceStatus == 1) {
@@ -86,7 +86,10 @@ public class MenuController {
 
     @ResponseBody
     @RequestMapping(value = "/detail", method = RequestMethod.PUT, produces = {"application/json; charset=utf-8"})
-    public Object updateMenu(@RequestBody MenuDO menuDO) {
+    public Object updateMenu(@RequestBody MenuDO menuDO, HttpServletResponse resp) {
+        if (GlobalUtils.authenErrorStatus(resp)) {
+            return GlobalUtils.authenErrorManager(resp);
+        }
         FinalResult finalResultUpdate;
         int menuServiceStatus = menuService.updateMenu(menuDO);
         if (menuServiceStatus == 1) {
@@ -102,7 +105,10 @@ public class MenuController {
      */
     @ResponseBody
     @RequestMapping(value = "/detail/{menuId}", method = RequestMethod.GET, produces = {"application/json; charset=utf-8"})
-    public Object getMenu(@PathVariable("menuId") int menuId) {
+    public Object getMenu(@PathVariable("menuId") int menuId, HttpServletResponse resp) {
+        if (GlobalUtils.authenErrorStatus(resp)) {
+            return GlobalUtils.authenErrorManager(resp);
+        }
         FinalResult finalResult = new FinalResult<MenuDO>(
                 true,
                 menuService.getMenu(menuId),
@@ -117,7 +123,10 @@ public class MenuController {
      */
     @ResponseBody
     @RequestMapping(value = "/detail/{menuId}", method = RequestMethod.DELETE, produces = {"application/json; charset=utf-8"})
-    public Object removeMenu(@PathVariable("menuId") int menuId) {
+    public Object removeMenu(@PathVariable("menuId") int menuId, HttpServletResponse resp) {
+        if (GlobalUtils.authenErrorStatus(resp)) {
+            return GlobalUtils.authenErrorManager(resp);
+        }
         FinalResult finalResultUpdate;
         int menuServiceStatus = menuService.removeMenu(menuId);
         if (menuServiceStatus == 1) {
